Trojan/PHP.WebShell
This Trojan arrives on a system as a file dropped by other malware or as a file downloaded unknowingly by users when visiting malicious sites. It may be hosted on a website and run when a user accesses the said website.

This malicious script runs on web browsers, which may be installed on any operating system.